+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to linkhut will be documented in this file.
 ## [Unreleased]
 
 - Add more MIME types that can be compressed at-rest.
+- Introduce concept of subscriptions instead of relying on user types to determine Archiving eligibility
 
 ## [0.1.1] - 2026-03-10